# cli/openapi.json
{
    "openapi": "3.0.2",
    "info": {
        "title": "CHIP Test Harness",
        "version": "0.1.0"
    },
    "paths": {
...
        "/api/v1/operators/": {
            "get": {
                "tags": [
                    "operators"
                ],
                "summary": "Read Operators",
                "description": "Retrive list of operators.\n\nArgs:\n    skip (int, optional): Pagination offset. Defaults to 0.\n    limit (int, optional): max number of records to return. Defaults to 100.\n\nReturns:\n    List[Operator]: List of operators",
                "operationId": "read_operators_api_v1_operators__get",
                "parameters": [
                    {
                        "required": false,
                        "schema": {
                            "title": "Skip",
                            "type": "integer",
                            "default": 0
                        },
                        "name": "skip",
                        "in": "query"
                    },
                    {
                        "required": false,
                        "schema": {
                            "title": "Limit",
                            "type": "integer",
                            "default": 100
                        },
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "title": "Response Read Operators Api V1 Operators  Get",
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Operator"
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "operators"
                ],
                "summary": "Create Operator",
                "description": "Create new operator.\n\nArgs:\n    operator_in (OperatorCreate): Parameters for new operator.\n\nReturns:\n    Operator: newly created operator record",
                "operationId": "create_operator_api_v1_operators__post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OperatorCreate"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Operator"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/operators/{id}": {
            "get": {
                "tags": [
                    "operators"
                ],
                "summary": "Read Operator",
                "description": "Lookup operator by id.\n\nArgs:\n    id (int): operator id\n\nRaises:\n    HTTPException: if no operator exists for provided operator id\n\nReturns:\n    Operator: operator record",
                "operationId": "read_operator_api_v1_operators__id__get",
                "parameters": [
                    {
                        "required": true,
                        "schema": {
                            "title": "Id",
                            "type": "integer"
                        },
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Operator"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "operators"
                ],
                "summary": "Update Operator",
                "description": "Update an existing operator.\n\nArgs:\n    id (int): operator id\n    operator_in (schemas.OperatorUpdate): operators parameters to be updated\n\nRaises:\n    HTTPException: if no operator exists for provided operator id\n\nReturns:\n    Operator: updated operator record",
                "operationId": "update_operator_api_v1_operators__id__put",
                "parameters": [
                    {
                        "required": true,
                        "schema": {
                            "title": "Id",
                            "type": "integer"
                        },
                        "name": "id",
                        "in": "path"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OperatorUpdate"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Operator"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "operators"
                ],
                "summary": "Delete Operator",
                "description": "Lookup operator by id.\n\nArgs:\n    id (int): operator id\n\nRaises:\n    HTTPException: if no operator exists for provided operator id\n\nReturns:\n    Operator: operator record that was deleted",
                "operationId": "delete_operator_api_v1_operators__id__delete",
                "parameters": [
                    {
                        "required": true,
                        "schema": {
                            "title": "Id",
                            "type": "integer"
                        },
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Operator"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "HTTPValidationError": {
                "title": "HTTPValidationError",
                "type": "object",
                "properties": {
                    "detail": {
                        "title": "Detail",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        }
                    }
                }
            },
...
            "Operator": {
                "title": "Operator",
                "required": [
                    "name",
                    "id"
                ],
                "type": "object",
                "properties": {
                    "name": {
                        "title": "Name",
                        "type": "string"
                    },
                    "id": {
                        "title": "Id",
                        "type": "integer"
                    }
                },
                "description": "Default schema, used when return data to API clients"
            },
            "OperatorCreate": {
                "title": "OperatorCreate",
                "required": [
                    "name"
                ],
                "type": "object",
                "properties": {
                    "name": {
                        "title": "Name",
                        "type": "string"
                    }
                },
                "description": "Create schema.\n\nName is required for new Operators."
            },
            "OperatorUpdate": {
                "title": "OperatorUpdate",
                "type": "object",
                "properties": {
                    "name": {
                        "title": "Name",
                        "type": "string"
                    }
                },
                "description": "Update Schema.\n\nSame as the base schema, only name can be changed"
            },
...
            "ValidationError": {
                "title": "ValidationError",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "type": "object",
                "properties": {
                    "loc": {
                        "title": "Location",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "msg": {
                        "title": "Message",
                        "type": "string"
                    },
                    "type": {
                        "title": "Error Type",
                        "type": "string"
                    }
                }
            },
...
        }
    }
}